PhysicalDevice moved to separate file
